requirements.txt
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
binwalk==2.2.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.17
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
deluge==2.0.3
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
folder-color-common==0.0.86
future==0.18.2
GeoIP==1.3.2
gpg===1.13.1-unknown
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jsonschema==3.2.0
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
meson==0.53.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.17.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
opencv-python==4.2.0.34
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
powerline-status==2.7
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.6
pyglet==1.5.5
PyGObject==3.36.0
pygubu==0.9.8.6
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.6.8
pytesseract==0.3.4
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-libtorrent==1.1.13
python-xapp==1.8.1
python-xlib==0.23
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
rencode==1.0.6
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rubber==1.5.1
scipy==1.3.3
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
speedtest-cli==2.1.2
system-service==0.3
systemd-python==234
texttable==1.6.2
torbrowser-launcher==0.3.2
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
unity-tweak-tool==0.0.7
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
youtube-dl==2020.3.24
zipp==1.0.0
zope.interface==4.7.1
